+++ src/main/java/com/ohgiraffers/semiproject/manager/search/model/service/SearchUserService.java
@@ -5,12 +5,10 @@ import com.ohgiraffers.semiproject.manager.search.model.dto.CartDTO;
 import com.ohgiraffers.semiproject.manager.search.model.dto.ProjectDTO;
 import com.ohgiraffers.semiproject.manager.search.model.dto.UserDTO;
 import com.ohgiraffers.semiproject.manager.search.model.dto.UserReportHistoryDTO;
-import org.apache.catalina.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Map;
 
 @Service
 public class SearchUserService {
@@ -25,24 +23,24 @@ public class SearchUserService {
     }
 
 
-    public List<CartDTO> userBuy(int no) {
+    public List<CartDTO> userBuy(long no) {
         List<CartDTO> cartDTOS = Mapper.userBuy(no);
         return cartDTOS;
     }
 
-    public List<UserReportHistoryDTO> userReport(int no) {
+    public List<UserReportHistoryDTO> userReport(long no) {
         List<UserReportHistoryDTO> userReportHistoryDTOS = Mapper.userReport(no);
 
         return userReportHistoryDTOS;
     }
 
-    public List<ProjectDTO> userFundingProject(int no) {
+    public List<ProjectDTO> userFundingProject(long no) {
         List<ProjectDTO> projectDTOS = Mapper.userFundingProject(no);
 
         return projectDTOS;
     }
 
-    public UserDTO findOneUser(int no) {
+    public UserDTO findOneUser(long no) {
         UserDTO userDTO = Mapper.findOneUser(no);
 
         return userDTO;
